m4ra_hampi

A sample street network from the township of Hampi, Karnataka, India. The same as the dodgr::hampi dataset, but in sc rather than sf format.

Format

A silicate sc representation of the street network of Hampi.

Note

Can be re-created with the following commands.

Examples

if (FALSE) { # \dontrun{
hampi <- dodgr_streetnet_sc ("hampi india")
} # }