Return a unique hash for the input network.

This can be used to identify cached versions of a given network, in particular to determine whether or not weighted versions of a given network have been previously cached.

Usage

m4ra_network_hash(graph)

Arguments

graph

A silicate, "SC", format object containing network data used to generate weighted street networks.

Value

Single character value with unique hash of given network.

See also